fix: includes missing prop to alerts-position
